requirement.txt
pillow=9.1.0
pygame=2.1.2
